Functional but ugly
I'm not particularly inspired as of right now as to how I want this to look so I'm gonna put aesthetics on the back burner because I need to get some actual coding practice in. But it works and fulfills all the challenges and optional challenges listed on the freeCodeCamp website.